bruh

+++ src/runner.c
@@ -1185,7 +1185,7 @@ void Runner_drawGUI(Runner* runner, int32_t windowW, int32_t windowH, int32_t ta
 
     if (runner->fpsRealFrameStartNanos != 0) {
         uint64_t elapsed = nowNanos() - runner->fpsRealFrameStartNanos;
-        if (elapsed > 0) runner->fpsReal = (double)1e9 / (double)elapsed;
+        if (elapsed > 0) runner->fpsReal = (double)1e9 / (double)(int64_t)elapsed;
     }
 }
